requirements.txt
boto >=2.0b4
argparse >=1.2.1
PyYAML
